+change int to size_t

+++ cplusplus_basic/memory/Memory.hpp
@@ -9,9 +9,9 @@ typedef void(*function_memory_not_enough)();
 _CPLUSPLUS_BASIC_LIBRARYSHARED_EXPORT function_memory_not_enough get_memory_not_enough();
 _CPLUSPLUS_BASIC_LIBRARYSHARED_EXPORT function_memory_not_enough set_memory_not_enough(function_memory_not_enough);
 _CPLUSPLUS_BASIC_LIBRARYSHARED_EXPORT void clean() noexcept(true);
-_CPLUSPLUS_BASIC_LIBRARYSHARED_EXPORT void * malloc(int arg) noexcept(true);
+_CPLUSPLUS_BASIC_LIBRARYSHARED_EXPORT void * malloc(size_t arg) noexcept(true);
 _CPLUSPLUS_BASIC_LIBRARYSHARED_EXPORT void free(void * arg) noexcept(true);
-_CPLUSPLUS_BASIC_LIBRARYSHARED_EXPORT int size(void * arg) noexcept(true);
+_CPLUSPLUS_BASIC_LIBRARYSHARED_EXPORT size_t size(void * arg) noexcept(true);
 _CPLUSPLUS_BASIC_LIBRARYSHARED_EXPORT int cookie_size() noexcept(true);
 _CPLUSPLUS_BASIC_LIBRARYSHARED_EXPORT void quick_exit(int=-2);
 inline void free(const void *arg) { memory::free(const_cast<void*>(arg)); }
